// src/client/api/generated/model/cliImageInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Spec Studio API
 * API local của Spec Studio. Lỗi chỉ trả mã (ErrorCode); FE tự dịch.
 * OpenAPI spec version: 1.0.0
 */
import type { CliImageInputMediaType } from './cliImageInputMediaType';

export interface CliImageInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  mediaType: CliImageInputMediaType;
  /**
     * @maxLength 7000000
     * @pattern ^[A-Za-z0-9+/]+={0,2}$
     */
  data: string;
}